requirements.txt
streamlit
opencv-python
numpy
paddleocr
easyocr
pytesseract
transformers
torch
Pillow
rapidocr-onnxruntime
